Update package names.
