+++ backend/src/main/com/tfttools/domain/Namable.java
@@ -1,0 +1,5 @@
+package main.com.tfttools.domain;
+
+public interface Namable {
+    String getDisplayName();
+}
