+++ backend/src/main/com/tfttools/dto/ChampionDTO.java
@@ -1,0 +1,17 @@
+package main.com.tfttools.dto;
+
+/**
+ * Sanitize champion object for data transfer from business logic to requestor
+ */
+public class ChampionDTO {
+
+    private final String displayName;
+
+    public ChampionDTO(String displayName) {
+        this.displayName = displayName;
+    }
+
+    public String getDisplayName() {
+        return displayName;
+    }
+}
