+++ backend/src/main/com/tfttools/prefixtrie/PrefixTrieUtils.java
@@ -1,0 +1,7 @@
+package main.com.tfttools.prefixtrie;
+
+public class PrefixTrieUtils {
+    public static String removePunctuation(String text) {
+        return text.toUpperCase().replaceAll("[,.'_\\-\\s]", "");
+    }
+}
