--- backend/src/main/java/com/tfttools/TFTToolsSpringBootApp.java
@@ -1,12 +1,0 @@
-package com.tfttools;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TFTToolsSpringBootApp
-{
-    public static void main(String[] args) {
-        SpringApplication.run(TFTToolsSpringBootApp.class, args);
-    }
-}
