--- backend/src/main/java/com/tfttools/config/SecurityFilterChainConfig.java
@@ -1,43 +1,0 @@
-package com.tfttools.config;
-
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.http.HttpMethod;
-
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
-import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.web.cors.CorsConfigurationSource;
-
-@Configuration
-@RequiredArgsConstructor
-@EnableWebSecurity
-public class SecurityFilterChainConfig
-{
-    private final CorsConfigurationSource corsConfigurationSource;
-
-    @Bean
-    public SecurityFilterChain securityFilterChain(HttpSecurity http)
-            throws Exception
-    {
-        http
-                .csrf(AbstractHttpConfigurer::disable)
-                .cors(httpSecurityCorsConfigurer -> httpSecurityCorsConfigurer.configurationSource(corsConfigurationSource))
-                .authorizeHttpRequests(authorizationManagerRequestMatcherRegistry ->
-                {
-                    authorizationManagerRequestMatcherRegistry.requestMatchers(
-                            HttpMethod.GET,
-                            "/units/filter",
-                            "/units/search"
-                    ).permitAll();
-                })
-                .sessionManagement(httpSecuritySessionManagementConfigurer -> httpSecuritySessionManagementConfigurer.sessionCreationPolicy(SessionCreationPolicy.STATELESS));
-        return http.build();
-    }
-
-}
