--- backend/src/main/java/com/tfttools/config/WebConfig.java
@@ -1,14 +1,0 @@
-package com.tfttools.config;
-
-import com.tfttools.converter.StringToChampionDTOConverter;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.format.FormatterRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class WebConfig implements WebMvcConfigurer {
-    @Override
-    public void addFormatters(FormatterRegistry registry) {
-        registry.addConverter(new StringToChampionDTOConverter());
-    }
-}
