--- backend/src/main/java/com/tfttools/converter/StringToChampionDTOConverter.java
@@ -1,13 +1,0 @@
-package com.tfttools.converter;
-
-import com.tfttools.dto.ChampionDTO;
-import org.springframework.core.convert.converter.Converter;
-import org.springframework.stereotype.Component;
-
-@Component
-public class StringToChampionDTOConverter implements Converter<String, ChampionDTO> {
-    @Override
-    public ChampionDTO convert(String source) {
-        return new ChampionDTO(source);
-    }
-}
