--- backend/src/main/java/com/tfttools/domain/Namable.java
@@ -1,5 +1,0 @@
-package com.tfttools.domain;
-
-public interface Namable {
-    String getDisplayName();
-}
