--- backend/src/main/java/com/tfttools/mapper/TraitMapper.java
@@ -1,19 +1,0 @@
-package com.tfttools.mapper;
-
-import com.tfttools.domain.Trait;
-import com.tfttools.dto.TraitDTO;
-import org.springframework.stereotype.Component;
-
-import java.util.function.Function;
-
-/**
- * Maps internal {@link Trait} object to {@link TraitDTO} object
- */
-@Component
-public class TraitMapper implements Function<Trait, TraitDTO> {
-
-    @Override
-    public TraitDTO apply(Trait trait) {
-        return new TraitDTO(trait.getDisplayName());
-    }
-}
