--- backend/src/main/java/com/tfttools/prefixtrie/PrefixTrieUtils.java
@@ -1,7 +1,0 @@
-package com.tfttools.prefixtrie;
-
-public class PrefixTrieUtils {
-    public static String removePunctuation(String text) {
-        return text.toUpperCase().replaceAll("[,.'_\\-\\s]", "");
-    }
-}
